### 2023-10-30 #6000 @editorialbot
[PRE REVIEW]: Fasten with Pipes
**Submitting author:** @lskatz (<a href="http://orcid.org/0000-0002-2533-9161">Lee Katz</a>)
**Repository:** https://github.com/lskatz/fasten
**Branch with paper.md** (empty if default branch):
**Version:** v0.7.2
