### 2023-11-01 #6000 @Kevin-Mattheus-Moerman
@lskatz thanks for this submission. I am the AEiC for this track, and here to help process the initial steps. For the moment please address the below in order for your submission to be considered for JOSS:

- [ ] In  your affiliations, please spell out `USA` as `United States of America`.
- [ ] The images (and therefore the font) in figure 1 are rather small. Please increase these, it seems feasible given that there is a lot of white space between them too.
- [ ] Your paper is not formatted according to our guidelines (and lacks the require "statement of need" section too). Please have a look here, https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain, and rewrite/reorganise your paper accordingly.
- [ ] I recommend that you extend your paper title as is usual for JOSS, e.g. some like the format `<project name>: <A sentence on what the software does>`
- [ ] The following sentence (the first one in the paper) seems a bit unclear to me (perhaps it is just me): `There are still many gaps in basic command line bioinformatics for standard file formats.`, do you mean something like: `There are still many gaps in basic command line tools for bioinformatics for handling of standard file formats.`?
