### 2023-11-01 #6000 @lskatz
Thank you for helping me on these points. For the resolution of the figure, would it help to separate it into 9 figures? When I combine them into a single image, I think my program is compressing the resolution. Or would having 9 figures make the manuscript too verbose?
