### 2023-11-01 873ccb8
trying 9 different figures instead of 1
