+++ paper/paper.md
@@ -44,23 +44,23 @@ Benchmarking was performed against other mainstream packages using `hyperfine` u
 
 Documentation, the container, and code are available at GitHub. Benchmarking results were graphed into Figures \label{fig:clean} - \label{fig:straighten}.
 
-![Trimming with a minimum quality score \label{fig:clean}](../tests/hyperfine/clean.json.png)
+![Trimming with a minimum quality score \label{fig:clean}](clean.json.png)
 
-![converting fastq to fasta \label{fig:convertToFasta}](../tests/hyperfine/convertToFasta.json.png)
+![converting fastq to fasta \label{fig:convertToFasta}](convertToFasta.json.png)
 
-![interleaving R1 and R2 reads \label{fig:interleave}](../tests/hyperfine/interleave.json.png)
+![interleaving R1 and R2 reads \label{fig:interleave}](interleave.json.png)
 
-![kmer counting \label{fig:kmer}](../tests/hyperfine/kmer.json.png)
+![kmer counting \label{fig:kmer}](kmer.json.png)
 
-![normalizing read depth using kmer coverage \label{fig:normalize}](../tests/hyperfine/normalize.json.png)
+![normalizing read depth using kmer coverage \label{fig:normalize}](normalize.json.png)
 
-![Searching for a sequence in a fastq file \label{fig:regex}](../tests/hyperfine/regex.json.png)
+![Searching for a sequence in a fastq file \label{fig:regex}](regex.json.png)
 
-![downsampling reads \label{fig:sample}](../tests/hyperfine/sample.json.png)
+![downsampling reads \label{fig:sample}](sample.json.png)
 
-![sorting fastq entries by either sequence or ID \label{fig:sort}](../tests/hyperfine/sort.json.png)
+![sorting fastq entries by either sequence or ID \label{fig:sort}](sort.json.png)
 
-![converting nonstandard fastq files to a format whose entries are four lines each, and selecting the first 100 \label{fig:straighten}](../tests/hyperfine/straighten.json.png)
+![converting nonstandard fastq files to a format whose entries are four lines each, and selecting the first 100 \label{fig:straighten}](straighten.json.png)
 
 ## Conclusions
 
