### 2023-11-01 #6000 @Kevin-Mattheus-Moerman
@lskatz I think a single merged figure is best. It may work if you simply have them nearly touching, so to remove the white space between them, and to then use the full page width. You also should be able to turn of compression for the software you are using. Anyway, give it a go and we'll look when you have the new draft.
