### 2023-11-01 c7e1935
one hi-res image in the paper
