### 2023-11-01 853347c
Joss editorial revision (#18)
* paper figures

* switch symlinks for hard links

* autoref correction

* m

* hi-res

* one hi-res image in the paper
