### 2023-11-02 #6000 @Kevin-Mattheus-Moerman
@lskatz this looks better indeed.  Below is some more feedback for you before we proceed.

- [ ] Although the paper now contains the required sections. I do think it would be better to rename the `Materials` section, for instance to `Implementation`. If there is any more implementation detail that you could add, that would not be doubling the documentation too much, then I would consider fleshing this section out a bit.
- [ ] In addition the `Results` section could perhaps be renamed for instance to something like `Documentation, testing, and benchmarking`. Next you can provide a little more detail on these aspects. For instance add some basic information on testing, e.g. what the tests cover, and what automated testing approaches are included. The benchmarking is interesting. Currently however, the reader has no idea what those images are about. So there is almost not point to have the image in the paper at the moment. It would be good to briefly explain the nature of the benchmarking, i.e. enough so that readers understand the nature of those graphs and the "performance" enhancements  you claim/present.
- [ ] Finally I would ask that you include a `CONTRIBUTING.md` file in your main repository folder, which contains required contributing guidelines. Next I would recommend adding a "How to contribute to Fasten" section in the README, and to link to the `CONTRIBUTING.md` file there. Here are some examples of such files: https://contributing.md/example/. If a dedicated file is too much for your project, you could instead simply have a longer section on this (detailing how you'd like folks to contribute to your project) in your README.

I will start looking for a handling editor but please consider the above.
