### 2023-11-02 e32d2ee
Bump rust docker version
