### 2023-11-02 05555c6
Try 1.73.0-bullseye for Docker
