### 2023-11-02 5259250
Bump fasten version on dockerfile
