### 2023-11-02 dd76dfa
Dockerfile: change to debian commands
