+++ Dockerfile
@@ -13,8 +13,8 @@ LABEL maintainer="Lee Katz"
 LABEL maintainer.email="gzu2@cdc.gov"
 LABEL maintainer2="John Phan"
 
-RUN apk update
-RUN apk add --no-cache \
+RUN apt-get update
+RUN apt-get install \
         ca-certificates \
         build-base \
         linux-headers \
@@ -26,6 +26,7 @@ RUN apk add --no-cache \
 RUN mkdir -p /usr/src/app \
     && cd /usr/src/app \
     && git clone https://github.com/lskatz/fasten \
+    && git checkout v${SOFTWARE_VER} \
     && cd /usr/src/app/fasten \
 RUN cd /usr/src/app/fasten && cargo build --release
 
