+++ .github/workflows/docker.yml
@@ -7,7 +7,7 @@ name: Create and publish a Docker image
 
 on:
   push:
-    branches: [master, main]
+    branches: [master, main, docker]
 
 env:
   REGISTRY: ghcr.io
