### 2023-11-02 441e5ce
Dockerfile: update apt-get packages for debian
