### 2023-11-02 c27891c
Dockerfile: Fix whitespace issue
