+++ Dockerfile
@@ -25,7 +25,7 @@ RUN mkdir -p /usr/src/app \
     && cd /usr/src/app \
     && git clone https://github.com/lskatz/fasten \
     && cd /usr/src/app/fasten \
-    && git checkout v${SOFTWARE_VER} \
+    && git checkout v${SOFTWARE_VER}
 RUN cd /usr/src/app/fasten && cargo build --release
 
 
