### 2023-11-02 7a3b99f
Dockerfile: cargo build debug target
