### 2023-11-02 229e0b3
Dockerfile: install bc
