### 2023-11-02 89f6f75
Docker fixes (#19)
* Dockerfile: change to debian commands

* docker.yml: run on docker branch too

* Dockerfile: update apt-get packages for debian

* Dockerfile: linux-headers-amd64

* Dockefile: apt-get install -y

* Dockerfile: Fix order

* Dockerfile: Fix whitespace issue

* Dockerfile: add tag label

* Docker.yml: update metadata-action

* Dockerfile: add unit tests

* Dockerfile: cargo build debug target

* Dockerfile: install bc
