+++ .github/workflows/docker.yml
@@ -7,7 +7,7 @@ name: Create and publish a Docker image
 
 on:
   push:
-    branches: [master, main]
+    branches: [master, main, docker]
 
 env:
   REGISTRY: ghcr.io
@@ -33,7 +33,7 @@ jobs:
 
       - name: Extract metadata (tags, labels) for Docker
         id: meta
-        uses: docker/metadata-action@98669ae865ea3cffbcbaa878cf57c20bbf1c6c38
+        uses: docker/metadata-action@v5
         with:
           images: ${{ env.REGISTRY }}/${{ env.IMAGE_NAME }}
 
