### 2023-11-02 #6000 @lskatz
Thank you! I your comments have made the manuscript better. I updated the manuscript's headers, and added onto the testing/benchmarking/documentation section, and added in the markdown files to guide in contributions.

Separately, I have separately asked @telatin if he would like to review this paper and he seems interested, and so I would like to put him forward as a possible reviewer.
