### 2023-11-07 #6030 @editorialbot
[REVIEW]: Fasten: a toolkit for streaming operations on fastq files
**Submitting author:** @lskatz (<a href="http://orcid.org/0000-0002-2533-9161">Lee Katz</a>)
**Repository:** https://github.com/lskatz/fasten
**Branch with paper.md** (empty if default branch):
**Version:** v0.8.1
