### 2023-11-08 #6030 @lskatz
@ reviewers I see there is a bullet point for `Performance` and so I just wanted to point you to the shell scripts buried in the repo that have the benchmarking: <https://github.com/lskatz/fasten/tree/master/tests>. The benchmarking scripts are benchmark_*.sh in that directory.
