### 2023-11-09 lskatz/fasten#20 @telatin
JOSS Review
Hello pips,
I open this issue to keep track of the review.

**Benchmark: version of the tools**
I started reproducing the benchmark and wanted to know if you could share the exact version of the programs used in your benchmark. I made an environment with the latest version of Fasten, SeqTK, Seqkit, SeqFu, fastx but maybe it's better to synchronise this (and to mention the versions used in the paper)

---
https://github.com/openjournals/joss-reviews/issues/6030
