### 2023-11-13 34ef373
Update bibliography and paper markdown (#21)
* Update paper.bib

* Update paper.md

* typos
* add SeqFu to the list, since it's in the benchmark
