+++ paper/paper.bib
@@ -9,19 +9,32 @@
 }
 
 @article{seqkit,
-    doi = {10.1371/journal.pone.0163962},
-    author = {Shen, Wei AND Le, Shuai AND Li, Yan AND Hu, Fuquan},
-    journal = {PLOS ONE},
-    publisher = {Public Library of Science},
-    title = {SeqKit: A Cross-Platform and Ultrafast Toolkit for FASTA/Q File Manipulation},
-    year = {2016},
-    month = {10},
-    volume = {11},
-    url = {https://doi.org/10.1371/journal.pone.0163962},
-    pages = {1-10},
-    abstract = {FASTA and FASTQ are basic and ubiquitous formats for storing nucleotide and protein sequences. Common manipulations of FASTA/Q file include converting, searching, filtering, deduplication, splitting, shuffling, and sampling. Existing tools only implement some of these manipulations, and not particularly efficiently, and some are only available for certain operating systems. Furthermore, the complicated installation process of required packages and running environments can render these programs less user friendly. This paper describes a cross-platform ultrafast comprehensive toolkit for FASTA/Q processing. SeqKit provides executable binary files for all major operating systems, including Windows, Linux, and Mac OSX, and can be directly used without any dependencies or pre-configurations. SeqKit demonstrates competitive performance in execution time and memory usage compared to similar tools. The efficiency and usability of SeqKit enable researchers to rapidly accomplish common FASTA/Q file manipulations. SeqKit is open source and available on Github at https://github.com/shenwei356/seqkit.},
-    number = {10},
+  doi = {10.1371/journal.pone.0163962},
+  url = {https://doi.org/10.1371/journal.pone.0163962},
+  year = {2016},
+  month = oct,
+  publisher = {Public Library of Science ({PLoS})},
+  volume = {11},
+  number = {10},
+  pages = {e0163962},
+  author = {Wei Shen and Shuai Le and Yan Li and Fuquan Hu},
+  editor = {Quan Zou},
+  title = {{SeqKit}: A Cross-Platform and Ultrafast Toolkit for {FASTA}/Q File Manipulation},
+  journal = {{PLOS} {ONE}}
+}
 
+@article{seqfu,
+  doi = {10.3390/bioengineering8050059},
+  url = {https://doi.org/10.3390/bioengineering8050059},
+  year = {2021},
+  month = may,
+  publisher = {{MDPI} {AG}},
+  volume = {8},
+  number = {5},
+  pages = {59},
+  author = {Andrea Telatin and Piero Fariselli and Giovanni Birolo},
+  title = {{SeqFu}: A Suite of Utilities for the Robust and Reproducible Manipulation of Sequence Files},
+  journal = {Bioengineering}
 }
 
 @software{seqtk,
