+++ paper/paper.md
@@ -22,9 +22,9 @@ bibliography: paper.bib
 ## Statement of need
 
 There are still many gaps in basic command line tools for the handling of standard file formats in the field of bioinformatics.
-Bioinformaticians have been able to use many tools to manipulate sequence data files in the fastq format, such as `seqkit` [@seqkit], `seqtk` [@seqtk] or FASTX-Toolkit [@fastx].
-These tools only accept paired end (PE) sequence data when split into multiple files per sample.
-Additionally, these tools do not always allow for Unix-style pipe file control. Sometimes they require explicity input/output options instead of using `stdin` and `stdout`.
+Bioinformaticians have been able to use many tools to manipulate sequence data files in the fastq format, such as `seqkit` [@seqkit], `seqtk` [@seqtk], FASTX-Toolkit [@fastx], or `seqfu` [@seqfu].
+These tools only accept paired-end (PE) sequence data when split into multiple files per sample.
+Additionally, these tools do not always allow for Unix-style pipe file control. Sometimes they require explicitly input/output options instead of using `stdin` and `stdout`.
 However, some bioinformaticians prefer to combine PE data from a single sample into one file using the interleaved fastq file format, but this format is not always well supported in mainstream tools.
 Here, we provide Fasten to the community to address these needs.
 
@@ -46,14 +46,14 @@ Continuous integration was implemented in GitHub Actions for unit testing.
 Each executable is tested to make sure the expected output is obtained with each `git push` event.
 We also used GitHub Actions to automatically create a Docker container which is also available on the GitHub repo.
 
-![Benchmarks comparing fasten with other analagous tools. From left to right, then to bottom: Trimming with a minimum quality score; converting fastq to fasta; interleaving R1 and R2 reads; kmer counting; normalizing read depth using kmer coverage; Searching for a sequence in a fastq file; downsampling reads; sorting fastq entries by either sequence or ID; and converting nonstandard fastq files to a format whose entries are four lines each, and selecting the first 100.\label{fig:benchmarks}](benchmarks.png)
+![Benchmarks comparing fasten with other analogous tools. From left to right, then to bottom: Trimming with a minimum quality score; converting fastq to fasta; interleaving R1 and R2 reads; kmer counting; normalizing read depth using kmer coverage; Searching for a sequence in a fastq file; downsampling reads; sorting fastq entries by either sequence or ID; and converting nonstandard fastq files to a format whose entries are four lines each, and selecting the first 100.\label{fig:benchmarks}](benchmarks.png)
 
 ## Conclusions
 
 Fasten is a powerful manipulation suite for interleaved fastq files, written in Rust.
 We benchmarked Fasten on several categories.
 It has strengths as shown in Figure 1 but it does not occupy the fastest position in all cases.
-Its major strengths include its competetive speeds,
+Its major strengths include its competitive speeds,
 Unix-style pipes,
 paired-end handling,
 and the advantages afforded by the Rust language including documentation and stability.
