### 2023-11-18 lskatz/fasten#25 @bovee
Clarify README summary
I think the summary would help potential users understand the package better if "random operations" and "secure your analysis" were reworded/expanded to be a bit more clear.

This covers the "statement of need" for documentation in the JOSS checklist: https://github.com/openjournals/joss-reviews/issues/6030
