### 2023-11-18 lskatz/fasten#26 @bovee
JOSS Paper citations
Jellyfish and SeqFu could be added to the References list.

Covered under "Software Paper" in "References" in the JOSS checklist: https://github.com/openjournals/joss-reviews/issues/6030
