### 2023-11-27 lskatz/fasten#20 @lskatz
CC @bovee

* Fasten 0.7.1
* SeqTK 1.0-r31
* seqkit 2.5.1
* Seq-Fu 1.20.0
* fastx-toolkit 0.0.13

These are the versions I used, but I can update to the latest if/where you see the opportunity.
