### 2023-11-27 5562961
cargo install command
