### 2023-11-27 2519679
added jellyfish and bbtools citations
