+++ paper/paper.bib
@@ -52,3 +52,20 @@
     url = {http://hannonlab.cshl.edu/fastx_toolkit/index.html},
     year = {2014}
 }
+
+@article{marcais2012jellyfish,
+  title={Jellyfish: A fast k-mer counter},
+  author={Marcais, G and Kingsford, C},
+  journal={Tutorialis e Manuais},
+  volume={1},
+  pages={1--8},
+  year={2012}
+}
+
+@techreport{bushnell2014bbmap,
+  title={BBMap: a fast, accurate, splice-aware aligner},
+  author={Bushnell, Brian},
+  year={2014},
+  institution={Lawrence Berkeley National Lab.(LBNL), Berkeley, CA (United States)}
+}
+
