### 2023-11-27 lskatz/fasten#26 @lskatz
Added these citations and BBTools for the same reason, in 2ddda21..2519679.

Thank you for pointing that out.
