### 2024-01-13 lskatz/fasten#20 @bovee
Following up, I was able to reproduce (modulo seqfu) the performance figure with the following on a quad core Intel Core i7-7700HQ with 16 Gb RAM running Manjaro Linux (kernel 5.10.199):
 - Fasten v0.7.2
 - seqtk 1.3-r106
 - seqkit v2.5.1
 - FASTX Toolkit 0.0.14
 - BBMap 39.06
 - jellyfish 2.3.1

Thanks for the benchmark scripts which made this much easier. For documentation in the paper, it might be good to include the hardware you're running the benchmarks on too?
