### 2024-01-19 lskatz/fasten#20 @telatin
Hi Lee,
my apologies for abandoning the review, the hiatus was longer than I thought.

I started checking the tests (note that I made a Makefile to automate building and tests, see https://github.com/telatin/fasten/tree/patch-2, if you want I can make a PR to add it but it depends on your style.)

In adding missing tests I noticed that `fasten_inspect` seems to work in paired mode by default:

```bash
cat testdata/four_reads.fasta | ./target/release/fasten_inspect
```
gives:
```text
>read0 id-at:0 seq-length:100 seq-invalid-chars: id-plus:0 qual-invalid-chars: avg-qual:38.16 qual-length:100 read-pair:1
AAAGTGCTCTTAACTTGTCCCGCTCCACATCAGCGCGACATCAATCGACATTAAACCGAGTATCTTGTCAGCCTGGGGTGACGATGCGTCCCATTAAAGT
>read1
TCTCATAGGGCTAACGCCAGAATCCCTCCATCGGTGTTTAAGGTGCCAAACACGATGGACCAAGTTATAGGGCTGACTATTAAACTATGGGGCCGGAAAC
>read2 id-at:0 seq-length:100 seq-invalid-chars: id-plus:0 qual-invalid-chars: avg-qual:39.60 qual-length:100 read-pair:1
GTTAGGAATTGAGGCCCTGGGGTCTGTGACGTCCGCGTGTAATATGAAAAACCAATATAATCAACTGTCCCACGCAATTATAAGCCGTAGTTCAACTTCG
>read3
CACGCGTACGTAGCAATTCCAACCTCGTTTCCTCAGTCTTCTAAGGAGGAAATGTGAGGTGATGTGTCTTCGTGACCTTTGGGCTTAACTCTCATAGAGA
```

i.e. no insights printed for read1 and read3
