+++ tests/fasten_normalize.sh
@@ -1,0 +1,6 @@
+#!/usr/bin/env bash
+# Minimal test suite for fasten normalise (telatin 2024)
+
+THIS_SCRIPT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
+THIS_SCRIPT_NAME=$(basename "$0" | sed 's/\.sh//g')
+source "${THIS_SCRIPT_DIR}/test_functions.sh"
