### 2024-01-29 lskatz/fasten#20 @telatin
In a PR added some minimal tests #30
