### 2024-01-29 0838287
Add makefile to run build and tests
