### 2024-02-05 460933d
benchmark_sample: some refinements
