### 2024-02-05 dd076c1
fasten_combine: refined awk
