### 2024-02-05 6edfcc9
Makefile and README: update with target/doc
