+++ .github/workflows/basic.yml
@@ -16,6 +16,10 @@ jobs:
         with:
           command: build
           #args: --release --all-features
+      - build
+        run: |
+          make releasae
+          make debug
       - name: tests
         run: |
           for i in tests/fasten_*.sh; do echo $i; echo ===; bash $i; echo; done;
