### 2024-02-05 e41ca44
note on input/output formats
