### 2024-02-05 #6030 @lskatz
- v0.8.0
  - https://doi.org/10.5281/zenodo.10621532
  - The license in Zenodo is Creative Commons, but mine is MIT.
