### 2024-02-05 #6030 @lskatz
I am not sure how to resolve the last checkbox with Creative Commons vs MIT.
