### 2024-02-05 #6030 @kellyrowland
hmm, can you try editing the metadata? https://help.zenodo.org/faq/technical/
