### 2024-02-05 #6030 @kellyrowland
thanks, the licenses match now. it looks like the authorship is inconsistent, however - John Phan is listed as an author on the Zenodo page but not in the paper itself. could the manuscript be updated?
