### 2024-02-06 #6030 @lskatz
Thank you for pointing that out. I have added @jhphan and accordingly, I bumped the version to 0.8.1. The new DOI is https://zenodo.org/doi/10.5281/zenodo.10621531 at https://zenodo.org/records/10625426.
