### 2024-02-07 #6030 @Kevin-Mattheus-Moerman
@lskatz As AEiC on this track I'll now help process this work for acceptance. I have just checked this review, the paper, your repository, and the archive link. I only have the below points which require your attention:

- [ ] Please ensure that the archive title matches your paper title `Fasten: a toolkit for streaming operations on fastq files`
- [ ] Please ensure that the archive license matches your software license (MIT).
