### 2024-02-09 #6030 @lskatz
Hi, I apologize, but I had one more edit on the proof. When I added @jhphan, I should have removed him in acknowledgements. I changed that here: https://github.com/lskatz/fasten/commit/6d37856157556c7e38b1c989cc0b3de1887046fd

I have edited the title accordingly

I have changed the license so that it matches the software license (MIT).
